# src/main/java/com/itdev/finalproject/dto/read/UserReadDto.java
package com.itdev.finalproject.dto.read;

import com.itdev.finalproject.database.entity.Role;

public record UserReadDto(
        Long id,
        String username,
        Role role,
        Integer age
) {
}